Added Cache Config

+++ src/main/java/com/fcr/trade/orders/model/OrderResponseDto.java
@@ -4,13 +4,12 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.math.BigDecimal;
-import java.util.Objects;
 
 @Getter
 @Setter
 public class OrderResponseDto {
     private String type;
-    private Integer totalNumberOfOrders;
+    private Long totalNumberOfOrders;
     private Long totalQuantity;
     private BigDecimal averagePrice;
     private String combinableOrder;
